+++ tools/build.ts
@@ -1,7 +1,8 @@
-import { BunPlatform_Args_Has } from '../src/lib/ericchase/BunPlatform_Args_Has.js';
+import { BunPlatform_Argv_Includes } from '../src/lib/ericchase/BunPlatform_Argv_Includes.js';
 import { Step_Dev_Format } from './core-dev/step/Step_Dev_Format.js';
 import { Step_Dev_Project_Update_Config } from './core-dev/step/Step_Dev_Project_Update_Config.js';
 import { Processor_HTML_Custom_Component_Processor } from './core-web/processor/Processor_HTML_Custom_Component_Processor.js';
+import { Processor_HTML_Remove_HotReload_On_Build } from './core-web/processor/Processor_HTML_Remove_HotReload_On_Build.js';
 import { Builder } from './core/Builder.js';
 import { Processor_Set_Writable } from './core/processor/Processor_Set_Writable.js';
 import { Processor_TypeScript_Generic_Bundler } from './core/processor/Processor_TypeScript_Generic_Bundler.js';
@@ -10,13 +11,11 @@ import { Step_FS_Clean_Directory } from './core/step/Step_FS_Clean_Directory.js'
 import { Step_Run_Self_Hosted_Server } from './lib-self-hosted-web-app/step/Step_Run_Self_Hosted_Server.js';
 import { Step_Self_Hosted_Server_Websocket_Reload } from './lib-self-hosted-web-app/step/Step_Self_Hosted_Server_Websocket_Reload.js';
 
-const SELF_HOSTED_SERVER_PORT = 54321;
-
 // If needed, add `cache` directory to the logger's file writer.
 // await AddLoggerOutputDirectory('cache');
 
 // Use command line arguments to set developer mode.
-if (BunPlatform_Args_Has('--dev')) {
+if (BunPlatform_Argv_Includes('--dev')) {
   Builder.SetMode(Builder.MODE.DEV);
 }
 // Set the logging verbosity
@@ -28,13 +27,12 @@ Builder.SetStartUpSteps(
   Step_Bun_Run({ cmd: ['bun', 'update', '--latest'], showlogs: false }),
   Step_Bun_Run({ cmd: ['bun', 'install'], showlogs: false }),
   Step_FS_Clean_Directory(Builder.Dir.Out),
-  Step_Dev_Format({ showlogs: false }),
   //
 );
 
 // These steps are run before each processing phase.
 Builder.SetBeforeProcessingSteps(
-  Step_Self_Hosted_Server_Websocket_Reload({ server_port: SELF_HOSTED_SERVER_PORT }),
+  Step_Self_Hosted_Server_Websocket_Reload(),
   //
 );
 
@@ -52,13 +50,14 @@ Builder.SetBeforeProcessingSteps(
 // The processors are run for every file that added them during every
 // processing phase.
 Builder.SetProcessorModules(
+  Processor_HTML_Remove_HotReload_On_Build(),
   // Process the HTML custom components.
   Processor_HTML_Custom_Component_Processor(),
   // Bundle the server.
   Processor_TypeScript_Generic_Bundler({ target: 'bun' }, { include_patterns: ['server.module.ts'], bundler_mode: 'module' }),
   // Bundle the IIFE scripts and module scripts.
-  Processor_TypeScript_Generic_Bundler({ define: () => ({ 'process.env.SERVERHOST': `127.0.0.1:${SELF_HOSTED_SERVER_PORT}` }) }, { bundler_mode: 'iife' }),
-  Processor_TypeScript_Generic_Bundler({ define: () => ({ 'process.env.SERVERHOST': `127.0.0.1:${SELF_HOSTED_SERVER_PORT}` }) }, { exclude_patterns: ['server.module.ts'], bundler_mode: 'module' }),
+  Processor_TypeScript_Generic_Bundler({}, { bundler_mode: 'iife' }),
+  Processor_TypeScript_Generic_Bundler({}, { exclude_patterns: ['server.module.ts'], bundler_mode: 'module' }),
   // Write non-bundle files and non-library files.
   Processor_Set_Writable({ include_patterns: ['**/*'] }),
   //
@@ -69,11 +68,14 @@ Builder.SetAfterProcessingSteps(
   // During developer mode (see above), the server will start running with
   // hot-reloading enabled for any of your HTML files that have called the
   // `EnableHotReload();` function in a script.
-  Step_Run_Self_Hosted_Server({ server_port: SELF_HOSTED_SERVER_PORT }),
+  Step_Run_Self_Hosted_Server(),
   //
 );
 
 // These steps are run during the cleanup phase only.
-Builder.SetCleanUpSteps();
+Builder.SetCleanUpSteps(
+  Step_Dev_Format({ showlogs: false }),
+  //
+);
 
 await Builder.Start();