+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 2025-08-16
+Please refer to the `ts-library` changelog for library and build tool changes:
 
-Please refer to the `ts-library` changelog: https://github.com/ericchase-library/ts-library/blob/main/CHANGELOG.md
+- https://github.com/ericchase-library/ts-library/blob/main/CHANGELOG.md
+
+## 2025-08-19
+
+- `lib-self-hosted-web-app` steps now acquire server host through logs